Start uapi on kernel configurer also
