+++ iface/wg_configurer_kernel.go
@@ -4,6 +4,7 @@ package iface
 
 import (
 	"fmt"
+	"golang.zx2c4.com/wireguard/device"
 	"net"
 	"time"
 
@@ -13,13 +14,41 @@ import (
 )
 
 type wgKernelConfigurer struct {
+	device     *device.Device
 	deviceName string
+
+	uapiListener net.Listener
 }
 
-func newWGConfigurer(deviceName string) wgConfigurer {
-	return &wgKernelConfigurer{
+func newWGConfigurer(device *device.Device, deviceName string) wgConfigurer {
+	wgc := &wgKernelConfigurer{
+		device:     device,
 		deviceName: deviceName,
 	}
+	wgc.startUAPI()
+	return wgc
+}
+
+func (t *wgKernelConfigurer) startUAPI() {
+	var err error
+	t.uapiListener, err = openUAPI(t.deviceName)
+	if err != nil {
+		log.Errorf("failed to open uapi listener: %v", err)
+		return
+	}
+
+	go func(uapi net.Listener) {
+		for {
+			uapiConn, uapiErr := uapi.Accept()
+			if uapiErr != nil {
+				log.Traceln("uapi Accept failed with error: ", uapiErr)
+				return
+			}
+			go func() {
+				t.device.IpcHandle(uapiConn)
+			}()
+		}
+	}(t.uapiListener)
 }
 
 func (c *wgKernelConfigurer) configureInterface(privateKey string, port int) error {
